requirements.txt
# Windows定时任务管理器 - 依赖包

# GUI框架
PySide6>=6.5.0

# 任务调度
APScheduler>=3.10.0

# 编码检测
chardet>=5.1.0

# 日期处理
python-dateutil>=2.8.2

# 开发工具
black>=23.0.0
pytest>=7.0.0
pytest-qt>=4.0.0